pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ansipix"
version = "1.0.0"
authors = [
  { name="EdgeOfAssembly", email="haxbox2000@gmail.com" },
]
description = "A high-performance terminal media viewer for playing images, GIFs, and videos in 24-bit truecolor."
readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia :: Graphics :: Viewers",
    "Topic :: Multimedia :: Video :: Display",
    "Topic :: Terminals",
]
keywords = ["terminal", "tui", "video", "gif", "image", "ansi", "truecolor", "viewer"]
dependencies = [
    "opencv-python",
    "numpy",
    "Pillow",
]

[project.urls]
Homepage = "https://github.com/EdgeOfAssembly/ansipix"
"Bug Tracker" = "https://github.com/EdgeOfAssembly/ansipix/issues"

# This section defines the command-line entry point.
# It tells pip to create a script named "ansipix" that calls the "cli" function
# inside the "ansipix.cli" module.
[project.scripts]
ansipix = "ansipix.cli:cli"